refactor: modularize 4 large service files in parallel
Modularize large service files into focused modules:

SVGRecordSheetRenderer.ts (1411 → 118 lines):
- 7 modules: constants, template, armor, structure, equipment, criticals, canvas

RecordSheetService.ts (1138 → 179 lines):
- 6 modules: constants, types, mechTypeUtils, equipmentUtils, criticalSlotUtils, dataExtractors

UnitLoaderService.ts (1034 → 62 lines):
- 7 modules: types, type-guards, component-mappers, equipment-resolution, equipment-mapping, armor-calculations, unit-loader

EquipmentRegistry.ts (931 → 362 lines):
- 5 modules: aliasUtils.weapon, aliasUtils.ammunition, aliasUtils.static, aliasUtils.legacy, aliasUtils.misc

Total: 4514 lines reduced to 721 lines in aggregators, split across 25 focused modules.
All 5301 tests passing, backward compatibility maintained.
